+++ crates/miden-objects/src/lib.rs
@@ -79,7 +79,8 @@ pub mod assembly {
 }
 
 pub mod crypto {
-    pub use miden_crypto::{SequentialCommit, dsa, hash, merkle, rand, utils};
+    pub use miden_crypto::{SequentialCommit, dsa, hash, ies, merkle, rand, utils};
+    pub use crypto_box::{PublicKey as PublicEncryptionKey, SecretKey as SecretDecryptionKey};
 }
 
 pub mod utils {